player initiative redone
- New initiative calculations, but only for the player. Monsters still
use the old one.
- The new system is a lot more obscure, but more flexible as well.
- Now initiative calculation includes weapon speed, armor encumbrance
and critical dice rolls that can anticipate or delay the phase.
- Dexterity at 18 gives a bonus to the attack phase.
- Agility to 18 gives a bonus to the movement phase.

- Formula for initiative is:
100 + offset (- weapon speed, - encumbrance) / 20.

So for example the initiative skill is 65% the 1d100 roll is 40. The
offset is the skill value - dice roll. So 65 - 40 = 25 is the offset.
Let's say weapon speed is 10 and encumbrance is 5, the result is 100 +
(25 - 10 - 5) = 110 / 2 = 55. That 55 is the initiative value being
used.

If the initiative value falls under 1 because of weapon and armor
penalties, a full phase is skipped.