fiddling with UI
Ongoing effort to move text to larger font and update the UI

- top bar support for 9x16
- larger 16x16 toggle buttons for topbar and right panel
- updated the 1234 map modes topbar buttons and some other topbar text (not finished)
- fixed minor aligment issue for 16x16 font almost never used
- updated right panel size and structure, and updated & fixed mouse highlight (still uses small font)

This more or less restores everything that was already working in the previous smaller window size version.

The silly goal would be now to move to the 9x16 font the initiative list and the whole message log, and create a custom toggable & scrollable message log window.
Then redo the combat prompt interface also with 9x16 font and much better layout.