libtcod upgrade
I forgot that a while back I tried a new version of libtcod, and it compiled fine. There were a few problems. The font source not being read in the right order, messing with the font, text aligment not working, and only being able to use the precompiled library.

Now I compiled the last source locally, aligment is fixed, I manually mapped the font and everything is working as before.

In the process I had to spend 4 hours trying to track and fix a crash in one of the mapmodes that happened only certain quirky conditions. The weird part is that the same code with old libtcod worked flawlessly, compiled with the new library it crashes. This was eventually fixed by zeroing some variable that I suspect were turning negative. So the attempt to draw outside the map made it crash.

Switch to new libtcod causes two more dlls to be added that for some reason won't compile statically: linstdc++ libgcc

- bunch of print switched to printf, because it's deprecated and produced warnings. Those with special characters turned into putChar
- now renders in OpenGL instead of SDL1

Actually the new version of libtcod seems to have much worse performance, in the sample program. 550fps the older one in SDL1 and 3500fps in OpenGL.
The new version gives me 140fps SDL1, 100fps SDL2, and 1200fps OpenGL.

But since the old library only let me use SDL1, the move to the new one and OpenGL means performance has doubled anyway.

Everything else seems working as before.